# Conflicts:
#	EP.Web/src/js/components/header.jsx

+++ EP.API/Areas/Admin/Controllers/BlobManagerController.cs
@@ -1,5 +1,4 @@
 ﻿using EP.API.Areas.Admin.ViewModels.Blobs;
-using EP.API.Extensions;
 using EP.API.Filters;
 using EP.API.ViewModels.Errors;
 using EP.Data.Entities.Blobs;
@@ -7,17 +6,10 @@ using EP.Data.Paginations;
 using EP.Services.Blobs;
 using EP.Services.Constants;
 using EP.Services.Logs;
-using EP.Services.Utilities;
-using EP.Services;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Net.Http.Headers;
-using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 using System;
-using EP.Data.Extensions;
 
 namespace EP.API.Areas.Admin.Controllers
 {
@@ -25,25 +17,19 @@ namespace EP.API.Areas.Admin.Controllers
     {
         private readonly IBlobService _blobService;
         private readonly IActivityLogService _activityLogService;
-        private readonly string _webRootPath;
-        private readonly string _publicBlob;
 
         public BlobManagerController(
             IBlobService blobService,
-            IActivityLogService activityLogService,
-            IHostingEnvironment hostingEnvironment,
-            AppSettings appSettings)
+            IActivityLogService activityLogService)
         {
             _blobService = blobService;
             _activityLogService = activityLogService;
-            _webRootPath = hostingEnvironment.WebRootPath;
-            _publicBlob = appSettings.PublicBlob;
         }
 
-        [HttpGet]
-        public async Task<IPagedList<Blob>> Get([FromQuery]BlobSearchViewModel viewModel)
+        [HttpGet("ChildList")]
+        public async Task<IPagedList<Blob>> ChildList(BlobSearchViewModel viewModel)
         {
-            return await _blobService.GetPagedListAsync(viewModel.Ext, viewModel.Page, viewModel.Size);
+            return await _blobService.GetChildListAsync(viewModel.Id, viewModel.Page, viewModel.Size);
         }
 
         [HttpGet("{id}")]
@@ -51,9 +37,11 @@ namespace EP.API.Areas.Admin.Controllers
         {
             var entity = await _blobService.GetByIdAsync(id);
 
-            if (entity == null || !System.IO.File.Exists(entity.PhysicalPath))
+            if (!_blobService.IsFile(entity))
             {
-                return NotFound();
+                ModelState.AddModelError(nameof(id), $"The {id} is not a file.");
+
+                return BadRequest(new ApiError(ModelState));
             }
 
             Stream fileStream = new FileStream(entity.PhysicalPath, FileMode.Open);
@@ -61,53 +49,111 @@ namespace EP.API.Areas.Admin.Controllers
             return File(fileStream, entity.ContentType);
         }
 
-        [HttpPost, ValidateMimeMultipartContent]
-        public async Task<IActionResult> Post([FromForm]BlobViewModel viewModel)
+        [HttpPost("Directory"), ValidateViewModel]
+        public async Task<IActionResult> PostDirectory([FromBody]DirectoryViewModel viewModel)
         {
-            var parent = await _blobService.GetByIdAsync(viewModel.Parent);
+            var parentEntity = await _blobService.GetByIdAsync(viewModel.Parent);
 
-            if (parent == null)
+            if (parentEntity == null)
             {
-                var parentName = nameof(viewModel.Parent);
-                ModelState.AddModelError(parentName, $"{parentName} is invalid.");
+                ModelState.AddModelError(nameof(viewModel.Parent), $"The {viewModel.Parent} is invalid.");
 
                 return BadRequest(new ApiError(ModelState));
             }
 
-            if (viewModel.IsValidDirectory)
-            {
+            string directoryName = viewModel.Name.Trim();
 
+            if (await _blobService.IsExistence(viewModel.Parent, directoryName))
+            {
+                ModelState.AddModelError(nameof(viewModel.Name), $"{directoryName} is existed.");
+
+                return BadRequest(new ApiError(ModelState));
             }
 
-            // if (files == null || files.Length == 0)
-            // {
-            //     ModelState.AddModelError(nameof(files), "Files should not be empty.");
+            var entity = new Blob
+            {
+                Name = directoryName,
+                Parent = viewModel.Parent,
+                CreatedOn = DateTime.UtcNow
+            };
 
-            //     return BadRequest(new ApiError(ModelState));
-            // }
+            await _blobService.CreateAsync(entity);
 
-            // Blob entity;
-            // IList<string> ids = new List<string>();
+            return Created(nameof(Directory), entity.Id);
+        }
 
-            // foreach (var file in files)
-            // {
-            //     entity = BuildBlob(file);
-            //     var activityLog = GetCreatedActivityLog(entity.GetType(), entity);
+        //[HttpPost("File"), ValidateViewModel, ValidateMimeMultipartContent]
+        //public async Task<IActionResult> PostFile([FromForm]FileViewModel viewModel)
+        //{
+        //    if (viewModel.Files == null || files.Length == 0)
+        //    {
+        //        ModelState.AddModelError(nameof(files), "Files should not be empty.");
 
-            //     await Task.WhenAll(
-            //         _blobService.CreateAsync(entity),
-            //         file.SaveAsAsync(entity.PhysicalPath),
-            //         _activityLogService.CreateAsync(SystemKeyword.CreateBlob, activityLog));
+        //        return BadRequest(new ApiError(ModelState));
+        //    }
 
-            //     ids.Add(entity.Id);
-            // }
+        //    Blob entity;
+        //    IList<string> ids = new List<string>();
 
-            // return Created(nameof(Post), ids);
+        //    foreach (var file in files)
+        //    {
+        //        entity = BuildBlob(file);
+        //        var activityLog = GetCreatedActivityLog(entity.GetType(), entity);
+
+        //        await Task.WhenAll(
+        //            _blobService.CreateAsync(entity),
+        //            file.SaveAsAsync(entity.PhysicalPath),
+        //            _activityLogService.CreateAsync(SystemKeyword.CreateBlob, activityLog));
+
+        //        ids.Add(entity.Id);
+        //    }
+
+        //    return Created(nameof(Post), ids);
+        //}
+
+        [HttpPut("Directory/{id}"), ValidateViewModel]
+        public async Task<IActionResult> PutDirectory(string id, [FromBody]DirectoryViewModel viewModel)
+        {
+            var parentEntity = await _blobService.GetByIdAsync(viewModel.Parent);
+
+            if (parentEntity == null)
+            {
+                ModelState.AddModelError(nameof(viewModel.Parent), $"The {viewModel.Parent} is invalid.");
+
+                return BadRequest(new ApiError(ModelState));
+            }
+
+            string directoryName = viewModel.Name.Trim();
+
+            if (await _blobService.IsExistence(viewModel.Parent, directoryName))
+            {
+                ModelState.AddModelError(nameof(viewModel.Name), $"{directoryName} is existed.");
+
+                return BadRequest(new ApiError(ModelState));
+            }
+
+            var entity = new Blob
+            {
+                Id = id,
+                Name = directoryName,
+                Parent = viewModel.Parent
+            };
+
+            await _blobService.UpdateAsync(entity);
+
+            return NoContent();
         }
 
         [HttpDelete]
         public async Task<IActionResult> Delete(string id)
         {
+            if (await _blobService.HasChildren(id))
+            {
+                ModelState.AddModelError(nameof(id), $"The {id} has sub directories or files.");
+
+                return BadRequest(new ApiError(ModelState));
+            }
+
             var entity = await _blobService.DeleteAsync(id);
 
             if (entity == null)
@@ -115,7 +161,8 @@ namespace EP.API.Areas.Admin.Controllers
                 return NotFound();
             }
 
-            if (System.IO.File.Exists(entity.PhysicalPath))
+            if (_blobService.IsFile(entity) &&
+                System.IO.File.Exists(entity.PhysicalPath))
             {
                 System.IO.File.Delete(entity.PhysicalPath);
             }
@@ -126,66 +173,59 @@ namespace EP.API.Areas.Admin.Controllers
             return NoContent();
         }
 
-        // private async Task<bool> IsValidParentBlob(string parent)
+        // private Blob BuildBlob(IFormFile file, int randomSize = 7)
         // {
-        //     return parent.IsInvalidObjectId() ?
-        //         false :
-        //         await _blobService.GetBlobByIdAsync(parent);
+        //     string contentType = file.ContentType;
+        //     string firstMimeType = GetFirstMimeType(contentType);
+        //     string publicBlobPath = GetPublicBlobPath(_webRootPath, _publicBlob, firstMimeType);
+
+        //     string fileName = ContentDispositionHeaderValue.Parse(file.ContentDisposition).FileName.ToString().Trim('"');
+        //     string name = Path.GetFileNameWithoutExtension(fileName);
+        //     string extension = Path.GetExtension(fileName);
+        //     string newFileName = $"{name}_{RandomUtils.Numberic(randomSize)}{extension}";
+
+        //     return new Blob
+        //     {
+        //         Name = newFileName,
+        //         FileExtension = extension.ToLowerInvariant(),
+        //         ContentType = contentType,
+        //         VirtualPath = string.IsNullOrWhiteSpace(firstMimeType) ?
+        //             $"{_publicBlob}/{newFileName}" :
+        //             $"{_publicBlob}/{firstMimeType}/{newFileName}",
+        //         PhysicalPath = Path.Combine(publicBlobPath, newFileName),
+        //         CreatedOn = DateTime.UtcNow
+        //     };
         // }
 
-        private Blob BuildBlob(IFormFile file, int randomSize = 7)
-        {
-            string contentType = file.ContentType;
-            string firstMimeType = GetFirstMimeType(contentType);
-            string publicBlobPath = GetPublicBlobPath(_webRootPath ,_publicBlob, firstMimeType);
+        // private static string GetFirstMimeType(string contentType)
+        // {
+        //     if (string.IsNullOrWhiteSpace(contentType))
+        //     {
+        //         return null;
+        //     }
 
-            string fileName = ContentDispositionHeaderValue.Parse(file.ContentDisposition).FileName.ToString().Trim('"');
-            string name = Path.GetFileNameWithoutExtension(fileName);
-            string extension = Path.GetExtension(fileName);
-            string newFileName = $"{name}_{RandomUtils.Numberic(randomSize)}{extension}";
+        //     var types = contentType.Split('/', StringSplitOptions.RemoveEmptyEntries);
 
-            return new Blob
-            {
-                Name = newFileName,
-                FileExtension = extension.ToLowerInvariant(),
-                ContentType = contentType,
-                VirtualPath = string.IsNullOrWhiteSpace(firstMimeType) ?
-                    $"{_publicBlob}/{newFileName}" :
-                    $"{_publicBlob}/{firstMimeType}/{newFileName}",
-                PhysicalPath = Path.Combine(publicBlobPath, newFileName),
-                CreatedOn = DateTime.UtcNow
-            };
-        }
+        //     if (types.Length == 0 || string.IsNullOrWhiteSpace(types[0]))
+        //     {
+        //         return null;
+        //     }
 
-        private static string GetFirstMimeType(string contentType)
-        {
-            if (string.IsNullOrWhiteSpace(contentType))
-            {
-                return null;
-            }
+        //     return types[0];
+        // }
 
-            var types = contentType.Split('/', StringSplitOptions.RemoveEmptyEntries);
+        // private static string GetPublicBlobPath(string webRootPath, string publicBlob, string firstMimeType)
+        // {
+        //     string publicBlobPath = string.IsNullOrWhiteSpace(firstMimeType) ?
+        //         Path.Combine(webRootPath, publicBlob) :
+        //         Path.Combine(webRootPath, publicBlob, firstMimeType);
 
-            if (types.Length == 0 || string.IsNullOrWhiteSpace(types[0]))
-            {
-                return null;
-            }
+        //     if (!Directory.Exists(publicBlobPath))
+        //     {
+        //         Directory.CreateDirectory(publicBlobPath);
+        //     }
 
-            return types[0];
-        }
-
-        private static string GetPublicBlobPath(string webRootPath, string publicBlob, string firstMimeType)
-        {
-            string publicBlobPath = string.IsNullOrWhiteSpace(firstMimeType) ?
-                Path.Combine(webRootPath, publicBlob) :
-                Path.Combine(webRootPath, publicBlob, firstMimeType);
-
-            if (!Directory.Exists(publicBlobPath))
-            {
-                Directory.CreateDirectory(publicBlobPath);
-            }
-
-            return publicBlobPath;
-        }
+        //     return publicBlobPath;
+        // }
     }
 }
